Issue #56
ADD:
  - peer.go:getBloomFilter() Now handles getting a bloom filter from a
    remote node upon connection.

CHANGE:
  - lib/network/outgoing/requester has been moved into peer.go, as it's
    the only caller.
