+++ lib/chord/chord.go
@@ -1,0 +1,1 @@
+package chord
