+++ lib/network/outgoing/outgoing.go
@@ -1,0 +1,1 @@
+package outgoing
